+++ packages/adp-tooling/src/client/index.ts
@@ -1,3 +1,0 @@
-export * from './target-applications';
-export * from './target-systems';
-export * from './abap-provider';