+++ packages/adp-tooling/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/adp-tooling
 
+## 0.13.24
+
+### Patch Changes
+
+-   c81864a: fix: ADP Generator integration test mocking
+
+## 0.13.23
+
+### Patch Changes
+
+-   4f0eaae: feat: Enhance system and app prompt validations
+
 ## 0.13.22
 
 ### Patch Changes